+++ src/app/shared/components/index.ts
@@ -3,4 +3,4 @@ export * from './search-container/search-container.component';
 export * from './quick-action-buttons/quick-action-buttons.component';
 export * from './installation-options/installation-options.component';
 export * from './setting-control/setting-control.component';
-export * from './cron-input/cron-input.component';
+export * from '../remote-config/cron-input/cron-input.component';